+++ lib/rack/brotli/railtie.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Rack
+  module Brotli
+    class Railtie < ::Rails::Railtie
+      initializer "rack-brotli.middleware" do |app|
+        app.middleware.use(Rack::Brotli)
+      end
+    end
+  end
+end